gigantic leap in compilation
Adopt new strategy for compilation of Lesya->Olesya.

This new strategy recognizes that some terms in Olesya are special.
Namely, the implication terms. And it therefore assumes intrinsic
axions to deal with Lesya-style implications.

Big advantage of this strategy is that both `eval` and `axiom` are as
simple as they could get.